+++ halotools/mock_observables/pair_counters/marked_cpairs/marked_npairs_xy_z_engine.pyx
@@ -254,5 +254,7 @@ cdef f_type return_weighting_function(weight_func_id):
         return tweights
     elif weight_func_id==10:
         return exweights
+    elif weight_func_id==11:
+        return ratio_weights
     else:
         raise ValueError('marking function does not exist')