+++ halotools/mock_observables/pair_counters/marked_cpairs/marking_functions.pyx
@@ -145,3 +145,14 @@ cdef cnp.float64_t dotweights(cnp.float64_t* w1, cnp.float64_t* w2):
 
 
 
+cdef cnp.float64_t ratio_weights(cnp.float64_t* w1, cnp.float64_t* w2):
+    """
+    ratio weights
+    return w2[1] if w2[0]>w1[1]*w1[0], 0 otherwise
+    id: 11
+    expects length 2 arrays
+    """
+    if w2[0] > w1[0]*w1[1]:
+        return w2[1]
+    else:
+        return 0.0