+++ halotools/mock_observables/pairwise_velocities/velocity_marked_npairs_xy_z.py
@@ -18,7 +18,7 @@ __all__ = ('velocity_marked_npairs_xy_z', )
 
 def velocity_marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins, period=None,
         weights1=None, weights2=None,
-        weight_func_id=0, verbose=False, num_threads=1,
+        weight_func_id=1, verbose=False, num_threads=1,
         approx_cell1_size=None, approx_cell2_size=None):
     r"""
     Calculate the number of velocity weighted pairs
@@ -116,6 +116,36 @@ def velocity_marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins, period=None,
         2-D array of shape *(Nrp_bins,Npi_bins)* containing the weighted number counts
         of pairs. The exact values depend on ``weight_func_id``
         (which weighting function was chosen).
+
+    Examples
+    --------
+    For demonstration purposes we will work with
+    halos in the `~halotools.sim_manager.FakeSim`.
+
+    >>> from halotools.sim_manager import FakeSim
+    >>> halocat = FakeSim()
+
+    >>> x = halocat.halo_table['halo_x']
+    >>> y = halocat.halo_table['halo_y']
+    >>> z = halocat.halo_table['halo_z']
+
+    We transform our *x, y, z* points into the array shape used by the pair-counter by
+    taking the transpose of the result of `numpy.vstack`. This boilerplate transformation
+    is used throughout the `~halotools.mock_observables` sub-package:
+
+    >>> sample1 = np.vstack((x,y,z)).T
+
+    We will do the same to get a random set of velocities.
+
+    >>> vx = halocat.halo_table['halo_vx']
+    >>> vy = halocat.halo_table['halo_vy']
+    >>> vz = halocat.halo_table['halo_vz']
+    >>> velocities = np.vstack((x,y,z,vx,vy,vz)).T
+
+    >>> rp_bins = np.logspace(-2,-1,10)
+    >>> pi_bins = np.linspace(0, 10, 5)
+    >>> result = velocity_marked_npairs_xy_z(sample1, sample1, rp_bins, pi_bins, period=halocat.Lbox, weights1=velocities, weights2=velocities,)
+
     """
     result = _npairs_xy_z_process_args(sample1, sample2, rp_bins, pi_bins, period,
             verbose, num_threads, approx_cell1_size, approx_cell2_size)