+++ halotools/mock_observables/radial_profiles/radial_profile_3d.py
@@ -202,8 +202,8 @@ def radial_profile_3d(sample1, sample2, sample2_quantity,
     if period is None:
         x1in, y1in, z1in, x2in, y2in, z2in, period = (
             _enclose_in_box(
-                sample1[:, 0], sample1[:, 2], sample1[:, 2],
-                sample2[:, 0], sample2[:, 2], sample2[:, 2],
+                sample1[:, 0], sample1[:, 1], sample1[:, 2],
+                sample2[:, 0], sample2[:, 1], sample2[:, 2],
                 min_size=[max_rbins_absolute*3.0, max_rbins_absolute*3.0, max_rbins_absolute*3.0]))
     else:
         x1in = sample1[:, 0]