+++ halotools/mock_observables/tensor_calculations/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from .inertia_tensor import inertia_tensor_per_object
+from .tensor_derived_quantities import *