+++ halotools/mock_observables/tensor_calculations/engines/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from .inertia_tensor_3d_engine import inertia_tensor_per_object_engine