+++ halotools/mock_observables/tensor_calculations/engines/inertia_tensor_3d_engine.pyx
@@ -1,25 +1,22 @@
 """
 """
-from __future__ import (absolute_import, division, print_function, unicode_literals)
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
 cimport numpy as cnp
 cimport cython
 from libc.math cimport ceil
 
-from .marking_functions cimport *
-from .custom_marking_func cimport custom_func
+from ....utils import unsorting_indices
 
 __author__ = ('Andrew Hearin', 'Duncan Campbell')
-__all__ = ('marked_npairs_3d_engine', )
-
-ctypedef double (*f_type)(cnp.float64_t* w1, cnp.float64_t* w2)
+__all__ = ('inertia_tensor_per_object_engine', )
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.nonecheck(False)
-def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
-    weights1in, weights2in, weight_func_idin, rbins, cell1_tuple):
+def inertia_tensor_per_object_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in, weights2in,
+            rsmooth, cell1_tuple):
     """ Cython engine for counting pairs of points as a function of three-dimensional separation.
 
     Parameters
@@ -33,15 +30,6 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
     x2in, y2in, z2in : arrays
         Numpy arrays storing Cartesian coordinates of points in sample 2
 
-    weights1in : array
-        Numpy array storing the weights for points in sample 1
-
-    weights2in : array
-        Numpy array storing the weights for points in sample 2
-
-    weight_func_id : int, optional
-        weighting function integer ID.
-
     rbins : array
         Boundaries defining the bins in which pairs are counted.
 
@@ -53,16 +41,12 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
     Returns
     --------
     counts : array
-        Integer array of length len(rbins) giving the number of pairs
-        separated by a distance less than the corresponding entry of ``rbins``.
+        Integer array of shape (len(x1in), len(rbins)) giving the number of pairs
+        separated by a distance less than the corresponding entry of ``rbins``
+        for each point in ``x1in``.
 
     """
-    cdef int weight_func_id = weight_func_idin
-
-    cdef f_type wfunc
-    wfunc = return_weighting_function(weight_func_id)
-
-    cdef cnp.float64_t[:] rbins_squared = rbins*rbins
+    cdef cnp.float64_t rsmooth_squared = rsmooth*rsmooth
     cdef cnp.float64_t xperiod = double_mesh.xperiod
     cdef cnp.float64_t yperiod = double_mesh.yperiod
     cdef cnp.float64_t zperiod = double_mesh.zperiod
@@ -71,17 +55,25 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
     cdef int PBCs = double_mesh._PBCs
 
     cdef int Ncell1 = double_mesh.mesh1.ncells
-    cdef int num_rbins = len(rbins)
-    cdef cnp.float64_t[:] counts = np.zeros(num_rbins, dtype=np.float64)
 
-    cdef cnp.float64_t[:] x1 = np.ascontiguousarray(x1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:] y1 = np.ascontiguousarray(y1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:] z1 = np.ascontiguousarray(z1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:] x2 = np.ascontiguousarray(x2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:] y2 = np.ascontiguousarray(y2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:] z2 = np.ascontiguousarray(z2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
-    cdef cnp.float64_t[:, :] weights1 = np.ascontiguousarray(weights1in[double_mesh.mesh1.idx_sorted,:], dtype=np.float64)
-    cdef cnp.float64_t[:, :] weights2 = np.ascontiguousarray(weights2in[double_mesh.mesh2.idx_sorted,:], dtype=np.float64)
+    cdef cnp.float64_t[:] x1_sorted = np.ascontiguousarray(
+        x1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
+    cdef cnp.float64_t[:] y1_sorted = np.ascontiguousarray(
+        y1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
+    cdef cnp.float64_t[:] z1_sorted = np.ascontiguousarray(
+        z1in[double_mesh.mesh1.idx_sorted], dtype=np.float64)
+    cdef cnp.float64_t[:] x2_sorted = np.ascontiguousarray(
+        x2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
+    cdef cnp.float64_t[:] y2_sorted = np.ascontiguousarray(
+        y2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
+    cdef cnp.float64_t[:] z2_sorted = np.ascontiguousarray(
+        z2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
+
+    cdef cnp.float64_t[:] w2_sorted = np.ascontiguousarray(
+        weights2in[double_mesh.mesh2.idx_sorted], dtype=np.float64)
+
+    cdef int npts1 = len(x1_sorted)
+    cdef cnp.float64_t[:, :, :] inertia_tensor = np.zeros((npts1, 3, 3), dtype=np.float64)
 
     cdef cnp.int64_t icell1, icell2
     cdef cnp.int64_t[:] cell1_indices = np.ascontiguousarray(double_mesh.mesh1.cell_id_indices, dtype=np.int64)
@@ -113,27 +105,25 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
     cdef int num_y2_per_y1 = num_y2divs // num_y1divs
     cdef int num_z2_per_z1 = num_z2divs // num_z1divs
 
-    cdef cnp.float64_t x2shift, y2shift, z2shift, dx, dy, dz, dsq, weight
+    cdef cnp.float64_t x2shift, y2shift, z2shift, dx, dy, dz, dsq
     cdef cnp.float64_t x1tmp, y1tmp, z1tmp
     cdef int Ni, Nj, i, j, k, l
 
     cdef cnp.float64_t[:] x_icell1, x_icell2
     cdef cnp.float64_t[:] y_icell1, y_icell2
     cdef cnp.float64_t[:] z_icell1, z_icell2
-    cdef cnp.float64_t[:,:] w_icell1, w_icell2
+    cdef cnp.float64_t[:] w_icell2
+
+    cdef cnp.float64_t xx, yy, zz, xy, xz, yz, w2
+    cdef cnp.float64_t[:] sum_weights = np.zeros(len(x1_sorted), dtype=np.float64)
+
 
     for icell1 in range(first_cell1_element, last_cell1_element):
-
         ifirst1 = cell1_indices[icell1]
         ilast1 = cell1_indices[icell1+1]
-
-        #extract the points in cell1
-        x_icell1 = x1[ifirst1:ilast1]
-        y_icell1 = y1[ifirst1:ilast1]
-        z_icell1 = z1[ifirst1:ilast1]
-
-        #extract the weights in cell1
-        w_icell1 = weights1[ifirst1:ilast1,:]
+        x_icell1 = x1_sorted[ifirst1:ilast1]
+        y_icell1 = y1_sorted[ifirst1:ilast1]
+        z_icell1 = z1_sorted[ifirst1:ilast1]
 
         Ni = ilast1 - ifirst1
         if Ni > 0:
@@ -184,13 +174,10 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
                         ifirst2 = cell2_indices[icell2]
                         ilast2 = cell2_indices[icell2+1]
 
-                        #extract the points in cell2
-                        x_icell2 = x2[ifirst2:ilast2]
-                        y_icell2 = y2[ifirst2:ilast2]
-                        z_icell2 = z2[ifirst2:ilast2]
-
-                        #extract the weights in cell2
-                        w_icell2 = weights2[ifirst2:ilast2,:]
+                        x_icell2 = x2_sorted[ifirst2:ilast2]
+                        y_icell2 = y2_sorted[ifirst2:ilast2]
+                        z_icell2 = z2_sorted[ifirst2:ilast2]
+                        w_icell2 = w2_sorted[ifirst2:ilast2]
 
                         Nj = ilast2 - ifirst2
                         #loop over points in cell1 points
@@ -199,6 +186,14 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
                                 x1tmp = x_icell1[i] - x2shift
                                 y1tmp = y_icell1[i] - y2shift
                                 z1tmp = z_icell1[i] - z2shift
+
+                                xx = 0.
+                                yy = 0.
+                                zz = 0.
+                                xz = 0.
+                                xy = 0.
+                                yz = 0.
+
                                 #loop over points in cell2 points
                                 for j in range(0,Nj):
                                     #calculate the square distance
@@ -207,46 +202,39 @@ def marked_npairs_3d_engine(double_mesh, x1in, y1in, z1in, x2in, y2in, z2in,
                                     dz = z1tmp - z_icell2[j]
                                     dsq = dx*dx + dy*dy + dz*dz
 
-                                    weight = wfunc(&w_icell1[i,0], &w_icell2[j,0])
-                                    k = num_rbins-1
-                                    while dsq <= rbins_squared[k]:
-                                        counts[k] += weight
-                                        k=k-1
-                                        if k<0: break
+                                    w2 = w_icell2[j]
 
-    return np.array(counts)
+                                    if dsq < rsmooth_squared:
+                                        xx = dx*dx*w2
+                                        yy = dy*dy*w2
+                                        zz = dz*dz*w2
+                                        xy = dx*dy*w2
+                                        xz = dx*dz*w2
+                                        yz = dy*dz*w2
+
+                                        inertia_tensor[ifirst1 + i, 0, 0] += xx
+                                        inertia_tensor[ifirst1 + i, 1, 1] += yy
+                                        inertia_tensor[ifirst1 + i, 2, 2] += zz
+
+                                        inertia_tensor[ifirst1 + i, 0, 1] += xy
+                                        inertia_tensor[ifirst1 + i, 1, 0] += xy
+
+                                        inertia_tensor[ifirst1 + i, 0, 2] += xz
+                                        inertia_tensor[ifirst1 + i, 2, 0] += xz
+
+                                        inertia_tensor[ifirst1 + i, 1, 2] += yz
+                                        inertia_tensor[ifirst1 + i, 2, 1] += yz
+
+                                        sum_weights[ifirst1 + i] += w2
 
 
-cdef f_type return_weighting_function(weight_func_id):
-    """
-    returns a pointer to the user-specified weighting function.
-    """
+    # At this point, we have calculated our counts on the input arrays *after* sorting
+    # Since the order of counts matters in this calculation, we need to undo the sorting
+    sorted_tensor = np.array(inertia_tensor)
+    sum_weights_array = np.array(sum_weights)
+    idx_unsorted = unsorting_indices(double_mesh.mesh1.idx_sorted)
+    sum_weights_array = sum_weights_array[idx_unsorted]
+    return sorted_tensor[idx_unsorted, :, :], sum_weights_array
 
-    if weight_func_id==0:
-        return custom_func
-    elif weight_func_id==1:
-        return mweights
-    elif weight_func_id==2:
-        return sweights
-    elif weight_func_id==3:
-        return eqweights
-    elif weight_func_id==4:
-        return ineqweights
-    elif weight_func_id==5:
-        return gweights
-    elif weight_func_id==6:
-        return lweights
-    elif weight_func_id==7:
-        return tgweights
-    elif weight_func_id==8:
-        return tlweights
-    elif weight_func_id==9:
-        return tweights
-    elif weight_func_id==10:
-        return exweights
-    elif weight_func_id==11:
-        return ratio_weights
-    elif weight_func_id==12:
-        return dotweights
-    else:
-        raise ValueError('marking function does not exist')
+
+